set taskHandle[cID] differently for each channel

+++ code/app.py
@@ -1,6 +1,7 @@
 #!/Users/ssg/.pyenv/shims/python
 # -*- coding: utf-8 -*-
 import sys
+from functools import wraps
 import threading
 from os import listdir
 import platform
@@ -240,28 +241,28 @@ class RemApplication(QMainWindow):
         self.startPredictionButton.setCheckable(True)
         self.startPredictionButton.clicked.connect(self.startPrediction)
         self.startPredictionButton.resize(self.startPredictionButton.sizeHint())
-        self.startPredictionButton.move(5 * self.scale, 10 * self.scale)
+        self.startPredictionButton.move(int(5 * self.scale), int(10 * self.scale))
 
         quitButton = QtWidgets.QPushButton('Quit', self)
         quitButton.clicked.connect(QtCore.QCoreApplication.instance().quit)
         quitButton.resize(quitButton.sizeHint())
-        quitButton.move(85 * self.scale, 10 * self.scale)
+        quitButton.move(int(85 * self.scale), int(10 * self.scale))
 
         checkConnectionButton = QtWidgets.QPushButton('Test connection', self)
         checkConnectionButton.clicked.connect(self.check_connection)
         checkConnectionButton.resize(checkConnectionButton.sizeHint())
-        checkConnectionButton.move(180 * self.scale, 10 * self.scale)
+        checkConnectionButton.move(int(180 * self.scale), int(10 * self.scale))
 
         # change standardization of ch2
         self.nameLabel_terminal_label = QLabel(self)
         self.nameLabel_terminal_label.setText('Terminal:')
-        self.nameLabel_terminal_label.move(250 * self.scale, 35 * self.scale)
+        self.nameLabel_terminal_label.move(int(250 * self.scale), int(35 * self.scale))
         self.terminal_combobox = QtWidgets.QComboBox(self)
         self.terminal_combobox.addItem(self.terminal_str_diff)
         self.terminal_combobox.addItem(self.terminal_str_rse)
         self.terminal_combobox.addItem(self.terminal_str_nrse)
         # self.terminal_combobox.addItem(self.terminal_str_pseudo) # not available for NI DAQ USB-6210
-        self.terminal_combobox.move(310 * self.scale, 38 * self.scale)
+        self.terminal_combobox.move(int(310 * self.scale), int(38 * self.scale))
         self.terminal_combobox.resize(self.terminal_combobox.sizeHint())
         self.terminal_combobox.activated[str].connect(self.terminal_choice)
         self.terminal_combobox.setCurrentText(self.terminal_config)
@@ -269,47 +270,47 @@ class RemApplication(QMainWindow):
         # change model
         self.nameLabel_classifier = QLabel(self)
         self.nameLabel_classifier.setText('Model:')
-        self.nameLabel_classifier.move(300 * self.scale, 10 * self.scale)
+        self.nameLabel_classifier.move(int(300 * self.scale), int(10 * self.scale))
 
         self.classifier_combobox = QtWidgets.QComboBox(self)
         for classifier_type in self.classifier_types:
             self.classifier_combobox.addItem(classifier_type)
         self.classifier_combobox.resize(self.classifier_combobox.sizeHint())
-        self.classifier_combobox.move(340 * self.scale, 10 * self.scale)
+        self.classifier_combobox.move(int(340 * self.scale), int(10 * self.scale))
         self.classifier_combobox.activated[str].connect(self.classifier_choice)
 
         '''
         self.nameLabel_ch2_overwrite_label = QLabel(self)
         self.nameLabel_ch2_overwrite_label.setText('Overwrite by Ch2:')
-        self.nameLabel_ch2_overwrite_label.move(450 * self.scale, 5 * self.scale)
+        self.nameLabel_ch2_overwrite_label.move(int(450 * self.scale), int(5 * self.scale))
 
         self.overwriteOrNotButton = QtWidgets.QPushButton(self.label_notOverwrite, self)
         self.overwriteOrNotButton.clicked.connect(self.toggleOverwriteOrNot)
         self.overwriteOrNotButton.resize(self.overwriteOrNotButton.sizeHint())
         self.overwriteOrNotButton.update()
-        self.overwriteOrNotButton.move(530 * self.scale, 5 * self.scale)
+        self.overwriteOrNotButton.move(int(530 * self.scale), int(5 * self.scale))
         self.overwriteOrNotButton.setCheckable(True)
         '''
 
         # change standardization of eeg
         self.nameLabel_eeg_mode_label = QLabel(self)
         self.nameLabel_eeg_mode_label.setText('EEG mode:')
-        self.nameLabel_eeg_mode_label.move(610 * self.scale, 2 * self.scale)
+        self.nameLabel_eeg_mode_label.move(int(610 * self.scale), int(2 * self.scale))
         self.eeg_mode_combobox = QtWidgets.QComboBox(self)
         self.eeg_mode_combobox.addItem(self.eeg_mode_str_normalize)
         self.eeg_mode_combobox.addItem(self.eeg_mode_str_none)
-        self.eeg_mode_combobox.move(680 * self.scale, 5 * self.scale)
+        self.eeg_mode_combobox.move(int(680 * self.scale), int(5 * self.scale))
         self.eeg_mode_combobox.resize(self.eeg_mode_combobox.sizeHint())
         self.eeg_mode_combobox.activated[str].connect(self.eeg_mode_choice)
 
         # change standardization of ch2
         self.nameLabel_ch2_mode_label = QLabel(self)
         self.nameLabel_ch2_mode_label.setText('Ch2 mode:')
-        self.nameLabel_ch2_mode_label.move(610 * self.scale, 30 * self.scale)
+        self.nameLabel_ch2_mode_label.move(int(610 * self.scale), int(30 * self.scale))
         self.ch2_mode_combobox = QtWidgets.QComboBox(self)
         self.ch2_mode_combobox.addItem(self.ch2_mode_str_normalize)
         self.ch2_mode_combobox.addItem(self.ch2_mode_str_none)
-        self.ch2_mode_combobox.move(680 * self.scale, 33 * self.scale)
+        self.ch2_mode_combobox.move(int(680 * self.scale), int(33 * self.scale))
         self.ch2_mode_combobox.resize(self.ch2_mode_combobox.sizeHint())
         self.ch2_mode_combobox.activated[str].connect(self.ch2_mode_choice)
 
@@ -317,35 +318,35 @@ class RemApplication(QMainWindow):
         self.nameLabel_ch2_thresh = QLabel(self)
         self.nameLabel_ch2_thresh.setText('Overwrite threshold to W:')
         self.nameLabel_ch2_thresh.resize(self.nameLabel_ch2_thresh.sizeHint())
-        self.nameLabel_ch2_thresh.move(840 * self.scale, 15 * self.scale)
+        self.nameLabel_ch2_thresh.move(int(840 * self.scale), int(15 * self.scale))
         self.ch2_thresh = QLineEdit(self)
         self.ch2_thresh.setText(str(self.params.ch2_thresh_default))
-        self.ch2_thresh.move(1000 * self.scale, 15 * self.scale)
+        self.ch2_thresh.move(int(1000 * self.scale), int(15 * self.scale))
         self.ch2_thresh.resize(30, 20)
 
         # change usage of ch2
         self.nameLabel_ch2_usage_label = QLabel(self)
         self.nameLabel_ch2_usage_label.setText('Ch2 usage:')
-        self.nameLabel_ch2_usage_label.move(840 * self.scale, 60 * self.scale)
+        self.nameLabel_ch2_usage_label.move(int(840 * self.scale), int(60 * self.scale))
         self.ch2_usage_combobox = QtWidgets.QComboBox(self)
         self.ch2_usage_combobox.addItem(self.ch2_usage_str_dontshowCh2)
         self.ch2_usage_combobox.addItem(self.ch2_usage_str_showCh2)
         self.ch2_usage_combobox.addItem(self.ch2_usage_str_overwrite)
-        self.ch2_usage_combobox.move(910 * self.scale, 60 * self.scale)
+        self.ch2_usage_combobox.move(int(910 * self.scale), int(60 * self.scale))
         self.ch2_usage_combobox.resize(self.ch2_usage_combobox.sizeHint())
         self.ch2_usage_combobox.activated[str].connect(self.ch2_usage_choice)
         self.ch2_usage_combobox_setup()
 
         self.ylim_label_eeg = QLabel(self)
         self.ylim_label_eeg.setText('eeg y-max:')
-        self.ylim_label_eeg.move(40 * self.scale, 40 * self.scale)
+        self.ylim_label_eeg.move(int(40 * self.scale), int(40 * self.scale))
         self.ylim_label_eeg.resize(self.ylim_label_eeg.sizeHint())
         self.ylim_value_eeg_box = QLineEdit(self)
-        self.ylim_value_eeg_box.move(45 * self.scale, 60 * self.scale)
+        self.ylim_value_eeg_box.move(int(45 * self.scale), int(60 * self.scale))
         self.ylim_value_eeg_box.resize(40, 20)
 
         self.ylim_slider_eeg = QSlider(Qt.Vertical, self)
-        self.ylim_slider_eeg.move(10 * self.scale, 40 * self.scale)
+        self.ylim_slider_eeg.move(int(10 * self.scale), int(40 * self.scale))
         self.ylim_slider_eeg.resize(20, 80)
         self.ylim_slider_eeg.setMinimum(0)
         self.ylim_slider_eeg.setMaximum(100)
@@ -355,14 +356,14 @@ class RemApplication(QMainWindow):
 
         self.ylim_label_ch2 = QLabel(self)
         self.ylim_label_ch2.setText('ch2 y-max:')
-        self.ylim_label_ch2.move(160 * self.scale, 40 * self.scale)
+        self.ylim_label_ch2.move(int(160 * self.scale), int(40 * self.scale))
         self.ylim_label_ch2.resize(self.ylim_label_ch2.sizeHint())
         self.ylim_value_ch2_box = QLineEdit(self)
-        self.ylim_value_ch2_box.move(165 * self.scale, 60 * self.scale)
+        self.ylim_value_ch2_box.move(int(165 * self.scale), int(60 * self.scale))
         self.ylim_value_ch2_box.resize(40, 20)
 
         self.ylim_slider_ch2 = QSlider(Qt.Vertical, self)
-        self.ylim_slider_ch2.move(130 * self.scale, 40 * self.scale)
+        self.ylim_slider_ch2.move(int(130 * self.scale), int(40 * self.scale))
         self.ylim_slider_ch2.resize(20, 40)
         self.ylim_slider_ch2.setMinimum(0)
         self.ylim_slider_ch2.setMaximum(100)
@@ -371,7 +372,7 @@ class RemApplication(QMainWindow):
         self.ylim_slider_ch2.valueChanged.connect(self.ylim_change_ch2)
 
         self.ch2_thresh_slider = QSlider(Qt.Horizontal, self)
-        self.ch2_thresh_slider.move(1035 * self.scale, 20 * self.scale)
+        self.ch2_thresh_slider.move(int(1035 * self.scale), int(20 * self.scale))
         self.ch2_thresh_slider.resize(190, 20)
         self.ch2_thresh_slider.setMinimum(-8)
         self.ch2_thresh_slider.setMaximum(16)
@@ -384,17 +385,17 @@ class RemApplication(QMainWindow):
         self.label_graph_ch2.setFont(QtGui.QFont('SansSerif', 24))
         self.label_graph_ch2.setText('Epoch# : Prediction')
         self.label_graph_ch2.resize(self.label_graph_ch2.sizeHint())
-        self.label_graph_ch2.move(500 * self.scale, 50 * self.scale)
+        self.label_graph_ch2.move(int(500 * self.scale), int(50 * self.scale))
 
         self.label_graph_eeg = QLabel(self)
         self.label_graph_eeg.setFont(QtGui.QFont('SansSerif', 20))
         self.label_graph_eeg.setText('EEG')
-        self.label_graph_eeg.move(5 * self.scale, 205 * self.scale)
+        self.label_graph_eeg.move(int(5 * self.scale), int(205 * self.scale))
 
         self.label_graph_ch2 = QLabel(self)
         self.label_graph_ch2.setFont(QtGui.QFont('SansSerif', 20))
         self.label_graph_ch2.setText('Ch2')
-        self.label_graph_ch2.move(5 * self.scale, 405 * self.scale)
+        self.label_graph_ch2.move(int(5 * self.scale), int(405 * self.scale))
 
         self.font = QtGui.QFont()
         self.font.setPointSize(18)
@@ -411,13 +412,13 @@ class RemApplication(QMainWindow):
             self.listOfPredictionResults.append(PredictionResultLabel(self))
             predXLoc = (graphID * 300) + 125
             predYLoc = 90
-            self.listOfPredictionResults[graphID].move(predXLoc * self.scale, predYLoc * self.scale)
+            self.listOfPredictionResults[graphID].move(int(predXLoc * self.scale), int(predYLoc * self.scale))
 
             xLoc = (graphID * 300) + 50
             for chanID in range(2):
                 yLoc = (chanID * 200) + 120
-                self.listOfGraphs[chanID].append(DynamicGraphCanvas(self, width = 3 * self.scale, height = 2 * self.scale, dpi=100))
-                self.listOfGraphs[chanID][graphID].move(xLoc * self.scale, yLoc * self.scale)
+                self.listOfGraphs[chanID].append(DynamicGraphCanvas(self, width=3*self.scale, height=2*self.scale, dpi=100))
+                self.listOfGraphs[chanID][graphID].move(int(xLoc * self.scale), int(yLoc * self.scale))
 
         self.setWindowTitle('Sleep stage classifier')
         xSize = self.graphNum * 310
